backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[hiredis]==5.2.0
arq==0.26.1
pydantic[email]==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
cryptography==43.0.0
aiosmtplib>=2.0,<4.0
jinja2==3.1.4
python-multipart==0.0.12
aiohttp==3.11.0
httpx==0.27.0
aiofiles==24.1.0
structlog==24.4.0
orjson==3.10.11
